% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{ca_facebook}
\alias{ca_facebook}
\title{get number of facebook likes}
\usage{
ca_facebook(
  page_id = Sys.getenv("FB_PAGE_ID"),
  access_token = Sys.getenv("FB_PAGE_TOKEN")
)
}
\arguments{
\item{page_id}{integer. facebook id of the page. Defaults to environment variable FB_PAGE_ID.}

\item{access_token}{character. page access token that never expires. see readme how to generate it. Defaults to environment variable FB_PAGE_TOKEN.}
}
\value{
integer. number of Facebook likes.
}
\description{
get number of facebook likes
}
