% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsletter.R
\name{ca_newsletter}
\alias{ca_newsletter}
\title{get subscriber count for CorrelAid newsletter}
\usage{
ca_newsletter(
  mc_key = Sys.getenv("MAILCHIMP_KEY"),
  mc_list_id = Sys.getenv("MAILCHIMP_LIST_ID")
)
}
\arguments{
\item{mc_key}{character. API key from mailchimp. Defaults to environment variable MAILCHIMP_KEY}

\item{mc_list_id}{character. id of mailchimp list / audience (extract this manually previously). Defaults to environment variable MAILCHIMP_LIST_ID}
}
\value{
tibble
}
\description{
c.f. https://mailchimp.com/developer/api/marketing/lists/
}
