% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{ca_slack}
\alias{ca_slack}
\title{get number of slack members}
\usage{
ca_slack(slackr_path = ".slackr")
}
\arguments{
\item{slackr_path}{character. path to .slackr file. defaults to ".slackr"}
}
\value{
integer. number of CorrelAid slack members

tibble
}
\description{
get number of slack members
}
