% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{ca_twitter}
\alias{ca_twitter}
\title{get follower count for CorrelAid twitter account}
\usage{
ca_twitter(user = Sys.getenv("TWITTER_USER"))
}
\arguments{
\item{user}{character. Twitter user to get follower count for. Defaults to Sys.getenv("TWITTER_USER")}
}
\value{
integer. follower count
}
\description{
get follower count for CorrelAid twitter account
}
