% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsletter.R
\name{mailchimp_lists}
\alias{mailchimp_lists}
\title{helper function to get id and name of all your mailchimp lists / audiences}
\usage{
mailchimp_lists(
  mc_key = Sys.getenv("MAILCHIMP_KEY"),
  mc_list_id = Sys.getenv("MAILCHIMP_LIST_ID")
)
}
\arguments{
\item{mc_key}{character. API key from mailchimp. Defaults to environment variable MAILCHIMP_KEY}
}
\value{
list. list of all mailchimp audiences / lists including their name and id.
}
\description{
c.f. https://mailchimp.com/developer/api/marketing/lists/
}
