% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsletter.R
\name{split_api_key}
\alias{split_api_key}
\title{helper function to split mailchimp api key into its two components}
\usage{
split_api_key(mc_key)
}
\description{
helper function to split mailchimp api key into its two components
}
